LPS-82637 Use simpler splitting strategy
Due to Google's quotas, big documents could exhaust the allowed number
of requests per day. To alleviate this problem, we won't send the
whole document, but only the first MB of text. This should be enough
to get a good set of tags, and will only requier a request per
document.